Refactor(Backend): Resolve monolith by modularizing main.py and eliminating in-memory state
