+++ backend/database.py
@@ -1,0 +1,37 @@
+import os
+from pathlib import Path
+from dotenv import load_dotenv
+
+env_path = Path(__file__).parent / '.env'
+load_dotenv(dotenv_path=env_path)
+
+try:
+    from supabase import create_client, Client
+    url = os.environ.get("SUPABASE_URL")
+    key = os.environ.get("SUPABASE_SERVICE_KEY")
+    if not url or not key:
+        print("[ERROR] SUPABASE_URL or SUPABASE_SERVICE_KEY not set in backend/.env")
+        supabase = None
+    else:
+        supabase = create_client(url, key)
+except (ImportError, Exception) as e:
+    print(f"[WARNING] Supabase initialization failed: {e}")
+    supabase = None
+
+def get_system_settings(company_id: str) -> dict:
+    defaults = {
+        "ai_confidence_threshold": 0.80,
+        "duplicate_sensitivity": 0.85,
+        "enable_auto_resolve": False
+    }
+    if not supabase or not company_id:
+        return defaults
+    try:
+        res = supabase.table("system_settings").select(
+            "ai_confidence_threshold, duplicate_sensitivity, enable_auto_resolve"
+        ).eq("company_id", company_id).single().execute()
+        if res.data:
+            return {**defaults, **res.data}
+    except Exception as e:
+        print(f"[WARNING] Could not fetch system_settings for company_id={company_id}: {e}")
+    return defaults
