+++ backend/routers/admin.py
@@ -1,0 +1,42 @@
+from fastapi import APIRouter
+from backend.database import supabase
+
+router = APIRouter(prefix="/api", tags=["Admin"])
+
+@router.get("/profiles")
+async def api_get_profiles(role: str = None, status: str = None, limit: int = 50, offset: int = 0):
+    if not supabase: return []
+    query = supabase.table("profiles").select("*")
+    if role: query = query.eq("role", role)
+    if status: query = query.eq("status", status)
+    res = query.range(offset, offset + limit - 1).execute()
+    return res.data
+
+@router.patch("/profiles/{user_id}")
+async def api_update_profile(user_id: str, updates: dict):
+    if not supabase: return {}
+    res = supabase.table("profiles").update(updates).eq("id", user_id).execute()
+    return res.data[0] if res.data else {}
+
+@router.delete("/profiles/{user_id}")
+async def api_delete_profile(user_id: str):
+    if not supabase: return {"success": False}
+    supabase.table("profiles").delete().eq("id", user_id).execute()
+    try:
+        supabase.rpc('delete_user').execute()
+    except: pass
+    return {"success": True}
+
+@router.get("/companies")
+async def api_get_companies():
+    if not supabase: return []
+    res = supabase.table("companies").select("*").execute()
+    return res.data
+
+@router.get("/admin_requests")
+async def api_get_admin_requests(status: str = None, limit: int = 50, offset: int = 0):
+    if not supabase: return []
+    query = supabase.table("admin_requests").select("*")
+    if status: query = query.eq("status", status)
+    res = query.range(offset, offset + limit - 1).execute()
+    return res.data
